Don't write the same frequency to vfd, when it's already correct.
When output frequency is identical to the frequency we want, it is no
point to continue writing to the vfd. This wasn't working since I
compared an int to an float.

This commit moves the comma two places, to include the necessary values,
then cast the float to an int.