Move global variables to main().
There is one calculation that was inside a while loop, this has been
moved out, and is done one time, before the loop.
This is safe to do since the values acted upon has default values,
and can only be changed with flags at startup, after this they can't
change.